Fix handling of A in float negation and equality comparison, more flexible addition and subtraction, enable tests for asinf(), atanf(), cosf(), sinf(), sinhf().
git-svn-id: http://svn.code.sf.net/p/sdcc/code/trunk@8816 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ sdcc/src/stm8/gen.h
@@ -51,7 +51,7 @@ typedef struct asmop_byte
   union
   {
     reg_info *reg;    /* Register this byte is in. */
-    int stk; /* Stack offset for this byte. */
+    long int stk; /* Stack offset for this byte. */
   } byteu;
 } asmop_byte;
